Separate download_past_media logic and simplify formatter
Options removed from the formatter:
- Message ID.
  Not all media have an associated message ID (e.g. profile photos).
- Extension.
  This can be added automatically.

In addition the extension is prefixed with the media ID. This
ensures that the media will be downloaded only once even when
the formatting string changes.

This is only done for download_past media. The normal media queue
is currently broken (but Telegram is down so I can't test anyway).

Following commits should fix that.

Separating the logic into its own method allows to keep and always
use the same formatter dictionary, which increases the consistency.