Always use the new _download_media for media operations
This arises several problems. Currently, the code needed to call
this method is often duplicated around different places, which
needs cleaning up.

Another (bigger) issue is that currently we always rely on the
database to get display names from the database, but the entities
might not have been dumped yet.

This is still untested (Telegram still down)